mod_auth_kerb: fix warning on solaris
git-svn-id: svn+ssh://source-control.beyondtrust.com/sources/repo/bt-dev/trunk/Platform/src/linux@59826 9fc1c11e-11b2-4e14-94d8-6c5e05775969

+++ mod_auth_kerb/src/mod_auth_kerb.c
@@ -944,7 +944,7 @@ apply_httpd_mapping( request_rec *r,
 			if ( *tmp == '$' ) {
 			    tmp++;
 			    i = 0;
-			    while ( isdigit( *tmp ) && (i < (sizeof( arg )-1))  ) {
+			    while ( isdigit( (int) *tmp ) && (i < (sizeof( arg )-1))  ) {
 				arg[i++] = *tmp;
 				tmp++;
 			    }
